Bug 1175389 - Add a Bluetooth telnet command to notify upper stack of inquiry result.
